// src/app/Interfaces/loginRespDTO.ts
export class LoginRespDTO {
    dateofbirth: string;
    dateofjoining: string;
    designation: string;
    emailid: string;
    loginflag: string;
    message: string;
    organizationid: string;
    organizationname: string;
    token: string;
    userId: string;
    userfirstname: string;
    userlastname: string;
}